refactor: 유틸 함수를 별도의 폴더로 분리
