+++ docs/README.md
@@ -15,21 +15,22 @@
 ### ⚠️ 주의 사항 체크  
 - [x] @woowacourse/mission-utils의 Random 및 Console API를 사용하여 구현해야 한다.  
 - [x] 입출력/에러 문구 작성 시 요구 사항에 부합해야 한다.  
-  
-### ✍🏻 추가 구현 및 수정 사항  
-- [x] 유효성 검증 로직 별도의 함수로 분리
-- [x] 점수 관련 로직 별도의 함수로 분리
-- [x] 추측, 판정의 과정을 반복하는 로직을 별도의 메소드로 분리
-- [x] 인풋 관련 로직을 별도의 함수로 분리
-- [x] 정답을 App class의 속성으로 분리하여 전반적 로직 개선
-- [x] 게임 완료 로직을 별도의 함수로 분리
-- [x] pickRandomThreeNumbers를 별도의 파일로 분리
-- [x] 상수를 별도의 파일로 분리
-  
-### 🏭 구현 단계  
-- [x] 0. 요구 사항을 분석하여 기능 목록을 명확하게 정리한다.  
+    
+### ✍🏻 추가 구현 및 수정 사항    
+- [x] 유효성 검증 로직 별도의 함수로 분리  
+- [x] 점수 관련 로직 별도의 함수로 분리  
+- [x] 추측, 판정의 과정을 반복하는 로직을 별도의 메소드로 분리  
+- [x] 인풋 관련 로직을 별도의 함수로 분리  
+- [x] 정답을 App class의 속성으로 분리하여 전반적 로직 개선  
+- [x] 게임 완료 로직을 별도의 함수로 분리  
+- [x] pickRandomThreeNumbers를 별도의 파일로 분리  
+- [x] 상수를 별도의 파일로 분리  
+- [x] 유틸 함수를 별도의 폴더로 분리  
+   
+### 🏭 구현 단계   
+- [x] 0. 요구 사항을 분석하여 기능 목록을 명확하게 정리한다.   
 - [x] 1. 퀄리티 신경 쓰지 않고 일단 구현한다.  
 - [x] 2. 테스트를 통과하는지 확인한다.   
-- [ ] 3. 리팩토링을 하며 퀄리티를 높인다.  
-  - [ ] 가독성을 최대한 높인다.  
-  - [ ] Airbnb 컨벤션 문서를 보며 잘 지켰는지 검수한다.  
+- [x] 3. 리팩토링을 하며 퀄리티를 높인다.  
+  - [x] 가독성을 최대한 높인다.  
+  - [x] Airbnb 컨벤션 문서를 보며 잘 지켰는지 검수한다.
